[PP-47] splitting event components
